Reland "Rewrite previously excluded raw_ptr fields"
This is a reland of commit 8a98a9bad2a2f3fc9a7a8e34092e95b6d4705b2a

PatchSet1 is the original change
Later patchsets are additional fixes.

The initial change was reverted due to compilation failure detected on
`linux-asan-media-v8-arm-rel`. This is due to `off_t` type being
`long long` in this configuration and due to raw_ptr not allowing
pointer arithmetic with such types.

Original change's description:
> Rewrite previously excluded raw_ptr fields
>
> This rewrites pointers to byte buffers that were accidentally skipped by
> the rewriter. In addition to skipping string literals, the rewriter was
> wrongly skipping byte buffers.
>
> DanglingUntriaged-notes: Annotating pre-existing dangling pointers.
> Bug: 331840473
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5497157
> Owners-Override: Arthur Sonzogni <arthursonzogni@chromium.org>
> Commit-Queue: Ali Hijazi <ahijazi@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1306660}

Bug: 331840473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5575382
Owners-Override: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Ali Hijazi <ahijazi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1307468}

+++ media/formats/mp4/avc.cc
@@ -118,7 +118,8 @@ bool AVC::InsertParamSetsAnnexB(const AVCDecoderConfigurationRecord& avc_config,
 
   if (nalu.nal_unit_type == H264NALU::kAUD) {
     // Move insert point to just after the AUD.
-    config_insert_point += (nalu.data + nalu.size) - start;
+    config_insert_point +=
+        (nalu.data + base::checked_cast<size_t>(nalu.size)) - start;
   }
 
   // Clear |parser| and |start| since they aren't needed anymore and
