+++ media/gpu/v4l2/v4l2_vda_helpers.cc
@@ -266,7 +266,7 @@ bool H264InputBufferFragmentSplitter::AdvanceFrameFragment(const uint8_t* data,
         return true;
       }
     }
-    *endpos = (nalu.data + nalu.size) - data;
+    *endpos = (nalu.data + base::checked_cast<size_t>(nalu.size)) - data;
   }
   NOTREACHED_NORETURN();
 }
